Used cosmetic pens to draw lines, fixing problems with display lists.
The way we handled transformations meant that display lists that are
defined when a certain transformation matrix is in use, but used when
a different matrix is in effect, will be rendered incorrectly.
The solution is to untransform each display list and retransform it
with the current matrix. We use cosmetic pens to prevent lines from
being scaled inappropriately.


git-svn-id: https://svn.met.no/diana/trunk@2895 0ccb8871-b436-41a3-ae7d-e3ac59973202