Fixed bug in GeoToPhys()
A crash would occur if we copied a layer from the DrawingDialog to the
EditDrawingDialog without first showing any layer in the underlay by
pushing the Apply button in the DrawingDialog. In that case, the plotRect
variable would still be uninitialized The problem was that the
plotRect member would still be uninitialized upon calling GeoToPhys().


git-svn-id: https://svn.met.no/diana/trunk@4409 0ccb8871-b436-41a3-ae7d-e3ac59973202