Re-enabled basic undo/redo in new drawing tool
Basic undo/redo for the new drawing tool is now available
from the Edit Drawing Dialog (Shift+Control+B).

NOTE: Changes to the layer structure/state are still not supported
as undoable operations. For example, undoing an item removal will insert
the item back into the topmost of the currently selected layers, regardless
of which layer the item originally belonged to.

git-svn-id: https://svn.met.no/diana/trunk@4435 0ccb8871-b436-41a3-ae7d-e3ac59973202