Support colour and alpha style props for text items
Summary:
The colour and alpha style properties of a text item may
now be interactively modified. This goes for the text itself,
the background area, and the border.

Test Plan:
Observere that the colour and alpha style properties
are interactively modifiable via the 'Edit Style' context menu for
the text item.

Reviewers: davidb

Reviewed By: davidb

Differential Revision: https://phab.met.no/D55