Fixed bug: don't update props when closing a read-only props dialog
Summary:
The text value of a text item would be set to an empty string when
closing the property dialog with the Ok button (instead of the Cancel
button). Since the properties are read-only anyway, the problem is
automatically solved by providing just an Close button which is
connected to the QDialog::reject() slot (just like the Cancel button).
NOTE: We still don't know why the text value was set to zero when the QDialog::accept() slot was called, but it doesn't really matter now.

Test Plan:
Open the properties dialog of a text item and observe that the
text value is left unchanged when closing the dialog with the Close button.

Reviewers: davidb

Reviewed By: davidb

Differential Revision: https://phab.met.no/D56