Made text change undoable
Summary: Modifying the text value of a text item is now undoable.

Test Plan:
Modify the text value of a text item by double-clicking
the item, or selecting 'Edit Text' from the context menu.
Observe that it is possible to undo the change (open the
Edit Drawing Dialog (Shift+Control+B) and click the undo button).

Reviewers: davidb

Reviewed By: davidb

Differential Revision: https://phab.met.no/D57