Ensured certain polyline decorations are always filled
Summary:
This fix explicitly fills certain polyline decorations in order for
them not to depend on the fill state set elsewhere (e.g. as a result
of changing the fill state of the map).

Test Plan:
Ensure that the map is not filled (i.e. showing the land contours only).
Create a polyline with decoration (e.g. a warm front) and verify that
the decorations are filled.

Reviewers: lisbethb

Reviewed By: lisbethb

Differential Revision: https://phab.met.no/D76