Don't auto-jump to selection mode
After placing a drawing item, the toolbar stays in this mode
until another mode is manually set. This way the user doesn't have
to use the Control modifier to achieve sticky mode when placing multiple
items of the same type.

The selection mode is now entered either by clicking on the button or
by pressing the Escape key.